test(auth): remove auth contract and update tests
